asdf

+++ CompetitionDataFinal/pickle-data.py
@@ -35,7 +35,7 @@ def main():
                 rating[attrIndex] = rating[attrIndex]/maxes[attrIndex]
 
         funcIn = np.array(allInputData).astype(np.float32)
-        funcOut = np.array([ b[2] for b in both]).astype(np.float32)
+        funcOut = np.array([ b[2] for b in both]).astype(np.int64)
 
         print(funcIn)
         print(funcIn.dtype)
